Move tox.ini to the project's root folder
This commit moves the tox.ini file to the project's
root folder. As a user I will like to use the pattern
from this template in my Travis, Jenkins, and Zuul CI/CD
automation in a similar way we use Makefiles.

The changes included in this commit are:

- Moving tox.ini to the project's root folder.
- Adjusting the README.md initial documentation.
- Including a main test_requirements.txt with the
testing dependencies.
- Including a .gitignore file excluding any file not
related to this template pattern.
- Refactoring of the tox.ini file to: crawl from the root folder,
not running any package distbuild, adjusting the linters that
don't do any recursive check to not require adding environments
when new modules or plugins are added.
- Adjusting the GitHub workflow to run all the tox environments
without passing any extra arguments.
- All the documented common errors should be addressed in this commit.

Closes: #1